fix(pr-execute): address 2 findings from triage of PR #3102
- phase7-module-cap-approaching-0 — extract CheckOnBehalfAuthorizedNode and
  EnsureOnBehalfParticipantExistsNode from vfd_role_guards.py (489/500 lines)
  into new on_behalf_guards.py (180 lines); vfd_role_guards.py now 350 lines;
  nodes/__init__.py exports both classes directly from on_behalf_guards
  (CS-18-001, BTND-07-004)
- phase9-notes-prm06-gap-0 — add PRM-06 assertion authority section to
  notes/participant-role-management.md; update relevant_packages frontmatter
  (NF-06-001, ADR-0084)
